finish the structure with error

+++ src/util/cfg_1TB.rs
@@ -10,7 +10,7 @@ pub const LUN_NUM:u32   = (1 << LN_BITS);
 pub const CH_NUM:u32    = (1 << CH_BITS);
 pub const PL_NUM:u32    = (1 << PL_BITS);
 pub const EP_NUM:u32    = (1 << EP_BITS);
-pub const BL_BITS_NUM:u32 = (32 - CH_BITS - EP_BITS - PL_BITS - LN_BITS - PG_BITS);
+pub const BL_BITS_NUM:u32 = (32 - CH_BITS - EP_BITS - PL_BITS - LN_BITS - PG_BITS) as u32;
 pub const SECTORS_PER_CELL:u32 = (1 << (PL_BITS + EP_BITS));  // 4x4 = 16 or 2x4 = 8
 
 
@@ -22,7 +22,6 @@ pub const CFG_NAND_PLANE_NUM:u32 = 4;
 pub const CFG_NAND_PAGE_SIZE:u32 = 16384;
 pub const CFG_NAND_PAGE_SPARE:u32 = 1260;
 pub const CFG_NAND_EP_SIZE:u32 = 4096;
-
 
 pub const NVME_QUART_PLANE 	:u16   	= PL_BITS << 0;
 pub const NVME_SINGLE_PLANE	:u16	= 0 << 0;
